Refactor the Node/rule interface, pt 2
Now, there are two distinct concepts: Nodes and Edges. Nodes are sources or
targets (source files, object files, executables, etc), and Edges are rules to
create targets from a set of sources. The built-in functions used in bfg files
now do two things: they add an Edge to the list of all edges (used to generate
the actual build files), and they return the Node that would result from the
Edge (possibly used with other edges).