Ensure that all build Nodes are fully-realized
Fully-realized build nodes have their true filename(s) generated upon
creation, instead of waiting until later to do so. This simplifies the
handling of generated files and reduces the need to keep track of the current
environment while writing rules for the build files. In the future, this could
also allow us to use different compilers on a per-target basis (instead of
per-language).