Emit two targets for Windows DLLs
This helps us to follow a simple rule: one target for one file. Doing so
allows users to have greater control over each file, e.g. allowing them to
install only the DLL, and not the import library. Note that the Make and Ninja
backends don't currently support rules with multiple targets (this is pending
support in those backends for MSVC).