Create a depfile for find_files() calls to check when to re-run bfg
This works by creating a depfile that lists every directory find_files() has
checked. If any of them have a newer mtime than the build file, the build
system will tell bfg to regenerate the build file. As a potential
optimization, we could make bfg's regeneration perform a more-thorough check
of whether the find_files() calls have changed, but that's probably only
necessary if the regeneration would otherwise be slow.