Support spaces in filenames (at least under Make and Ninja)
This commit kind of got out of hand due to the necessity to restructure the
Make backend. It's kind of tough to see what the changes are, but hopefully
everything will be ok. It works surprisingly well, actually. I wasn't even
sure it'd be possible to support all this under GNU Make!