Refactor the Command API to make it easier to specify commands
It's no longer necessary to explicitly supply the command's name when getting
the command's args in most cases. This removes some redundancy and makes it
possible to more-generically generate command variables in the build scripts.
Notably, the jvmoutput command now gets a variable; it used to be impossible
to do this because the compile rules didn't know there were actually two
commands being run.

The downside is that this requires a bit more plumbing to make sure that the
backends' writers know how to write out Command objects (usually by creating
a variable for them and referring to that variable).