Improve support for MinGW
This restructures how the cc-style builder parses library options, making it
cleaner and more-correct for edge cases like packages loaded for MinGW. Some
stuff still doesn't work right on CI for MinGW, but it's a lot closer. In
particular, MinGW still can't link to .lib files, which (I think?) is a bug
in bfg.

This patch also changes the Windows CI builds to use a shared version of
libogg and fixes tests to account for that.