Simplify the compile/link_options methods for packages
This removes the `output` argument to the above methods, since they weren't
terribly useful anymore, especially in a world with semantic options. In the
future, if we need anything similar to this, we can just create a semantic
option and rely on the fact that the `flags` methods for compilers/linkers
accept an `output` argument, allowing us to do any output file-specific
processing there.