Override the lang of input files when explicitly passing `lang` to build steps
This helps to make the behavior of build steps more obvious and also allows
using builders whose language differs from the "real" source language of the
input files (e.g. with Qt's `moc`). In addition, this commit adds the notion
of a "base" language for known languages, to be used with moc (whose base lang
is c++).