Improve how file objects are cloned
Now, when cloning a file object, subfiles are cloned as well, updating their
paths as appropriate. This is needed to allow pkg-config to work with versioned
shared libraries. Otherwise, when cloning a link library, the compiler would
get confused when trying to generate an rpath for the associated runtime file,
which wasn't updated to its installed location.