Add ability to guess C-family compilers based on env vars set for others in the family
This is particularly useful for internal bits that want to use a compiler in
the C family other than the main one used by the project. For example, a C++
project might implicitly use the C compiler when looking up packages. This
patch helps ensure that the same kind of compiler is used for both.